feat(programs): reintroduce Breadcrumbs component in ProgramDetailPage and remove it from ProgramDetailPageRoute for better navigation chore(program-card): remove console log for cleaner code refactor(program-detail-page): restructure breadcrumb logic to enhance clarity and maintainability
